Add files via upload
PYTHON CALCULATOR: To ease the calculations I have created a Pycalculator. The program begins with the display of logo and a welcome statement for increasing user experience. Then I have created functions for ADDITION, SUBTRACTION, MULTIPLICATION and DIVISION. The program asks the User to enter the digits and the type of operation to be performed .The final result will be displayed.